% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass-installation.R
\name{grass_version}
\alias{grass_version}
\title{Get the GRASS GIS version}
\usage{
grass_version(gisBase)
}
\arguments{
\item{gisBase}{A character string with the path to the GRASS installation.}
}
\value{
A list containing cumulative version strings for the major version,
major-minor version, and full version.
}
\description{
Get the GRASS GIS version
}
\keyword{internal}
