% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass-installation.R
\name{search_grass}
\alias{search_grass}
\title{Search for GRASS GIS installation}
\usage{
search_grass()
}
\value{
a character string with the path to the GRASS GIS installation
}
\description{
Search for GRASS GIS installation
}
\keyword{internal}
