% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup-environment.R
\name{set_addons_path}
\alias{set_addons_path}
\title{Set the path to GRASS add-ons directory}
\usage{
set_addons_path(addon_base = NULL, gv)
}
\arguments{
\item{addon_base}{path to the GRASS addons directory}
}
\value{
The configured add-on path invisibly, or \code{NULL} if the directory
does not exist.
}
\description{
Set the path to GRASS add-ons directory
}
\keyword{internal}
