% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup-environment.R
\name{set_home_path}
\alias{set_home_path}
\title{Get the home path environment variable}
\usage{
set_home_path(home = NULL)
}
\arguments{
\item{home}{path to the home directory}
}
\value{
invisible home path
}
\description{
Get the home path environment variable
}
\keyword{internal}
