% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup-environment.R
\name{set_path_variable}
\alias{set_path_variable}
\title{Prepend a path to an environment variable}
\usage{
set_path_variable(variable_name, install_path)
}
\arguments{
\item{variable_name}{Name of the environment variable.}

\item{install_path}{Path to prepend.}
}
\value{
The updated environment-variable value, invisibly.
}
\description{
Adds a path to a path-list environment variable unless that exact path is
already present.
}
\keyword{internal}
