% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup-environment.R
\name{setup_runtime_env_windows}
\alias{setup_runtime_env_windows}
\title{Set Windows runtime environment variables for GRASS GIS}
\usage{
setup_runtime_env_windows(gisBase, addon_base = NULL, gv)
}
\arguments{
\item{gisBase}{Path to the GRASS installation.}

\item{addon_base}{Optional path to the GRASS add-ons directory.}

\item{gv}{Cumulative GRASS version strings returned by \code{grass_version()}.}
}
\value{
The value of \code{OSGEO4W_ROOT}, invisibly.
}
\description{
Set Windows runtime environment variables for GRASS GIS
}
\keyword{internal}
