% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass-installation.R
\name{validate_gisbase}
\alias{validate_gisbase}
\title{Check that the GRASS installation directory is valid}
\usage{
validate_gisbase(gisBase)
}
\arguments{
\item{gisBase}{Path to the GRASS installation directory}
}
\description{
Check that the GRASS installation directory is valid
}
\keyword{internal}
